manager: Use Long for versionCode to support values exceeding Int.MAX_VALUE (#872)

+++ manager/app/src/main/java/com/sukisu/ultra/ui/util/module/LatestVersionInfo.kt
@@ -1,7 +1,7 @@
 package com.sukisu.ultra.ui.util.module
 
 data class LatestVersionInfo(
-    val versionCode: Int = 0,
+    val versionCode: Long = 0L,
     val downloadUrl: String = "",
     val changelog: String = ""
 )
